Crate iota_transaction_checks

Modules

deny

Functions

aggregate_authenticator_input_objects

A function to aggregate the checked authenticator input objects for multiple MoveAuthenticators into one CheckedInputObjects to be used for execution.

check_certificate_and_move_authenticator_input

A function to check the MoveAuthenticator inputs for execution and then for certificate execution. To be used instead of check_certificate_input when there is a Move authenticator present.

check_certificate_input

check_dev_inspect_input

WARNING! This should only be used for the dev-inspect transaction. This transaction type bypasses many of the normal object checks

check_move_authenticator_input_for_signing

A common function to check the MoveAuthenticator inputs for signing.

check_non_system_packages_to_be_published

Check package verification timeout

check_transaction_input

check_transaction_input_with_given_gas